+++ Assets/Scripts/PlanetLvl.cs
@@ -10,7 +10,7 @@ public class PlanetLvl : MonoBehaviour
     public int planetLevel = 0;
 
     // Make sure the amount of levels is equal or higher than r.levels.Length in the inspector (check JSON)
-    public int[] levels = new int[] { 10000, 100000, 250000, 500000, 750000, 1000000, 1500000 };
+    public ulong[] levels = new ulong[] { 10000, 100000, 250000, 500000, 750000, 1000000, 1500000 };
     private MapGenerator mg;
     private RegionLevels rl;
